took out all the private variables
