+++ src/test/java/com/boggle/BoggleLetterTest.java
@@ -1,3 +1,5 @@
+package com.boggle;
+
 import com.boggle.BoggleLetter;
 
 public class BoggleLetterTest {
